mm: Add more efficient map_mdl for CR3 init
Instead of calling g_mm->descriptors for mapping in the VMM's virt
addresses into its CR3, create a new function map_mdl that accesses the
m_virt_map directly. This prevents the need for re-creating a list of
the descriptors as is done with g_mm->descriptors (and hence saves many
unnecessary malloc/frees).