mm: Add support for 2MB and 1GB pages for vmm memory
Now that the driver can supply memory for the vmm's buddy allocators,
a more efficient method is needed for adding and mapping the memory
descriptors into the vmm (since the amount of memory can be many GBs).

To this end, add support for mapping 2MB and 1GB pages into the
vmm. The driver is modified to add memory descriptors in 4K, 2M, or
1G chunks. The memory manager then places each descriptor into the
appropriate virt/phys map (one map per granularity). When it
is time to map, 1GB mappings are used when available in the CPU,
otherwise 2MB and then 4KB mappings are used.

The logic for unmapping the vmm from the root vm is simplified to
support the different granularities: all 1G, 2M, and 4K mappings are
unmapped, in that order. Then the allowed mappings (such as the
root vm xenstore/console pages) are remapped in.