xen/memory: Remove map_page dead code
The hypercall page of guest VMs comes from the kernel, and the
kernel is mapped into EPT at the time the init_hypercall_page function
is run. Remove the call to map_page since it is not necessary.
Since that is the only caller of the two-argument map_page, remove
it from the xen_memory class.