xen/memory: Add support for large pages in guest VMs
The toolstack specifies an extent_order when is calls populate_physmap
during domain construction. This order is either 0 (for 4K), 9 (for 2M),
or 18 (for 1G). Prior to this change, only for 4K mappings were used.

In order to support VMs with larger memory requirements, add support
for mapping orders 9 and 18. Also handle XENMEM_decrease_reservations
that specify pages that lie on a large page by adding code to shatter
the mapping.

Note that since EPT is shared with the IOMMU, the IOMMU(s) of a domain
must support the large page sizes as well.